Code refactor
- renamed AuthenticationController to UserController
- created interface for controller
- moved the user exceptions into their own package
