+++ backend/src/main/java/ubb/courses/backend/controllers/exceptions/user/UserExceptionController.java
@@ -1,4 +1,4 @@
-package ubb.courses.backend.controllers.exceptions.authentication;
+package ubb.courses.backend.controllers.exceptions.user;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import ubb.courses.backend.controllers.responses.ApiResponse;
 
 @ControllerAdvice
-public class AuthenticationExceptionController {
+public class UserExceptionController {
     @ExceptionHandler(value = UsernameException.class)
     public ResponseEntity<Object> usernameException(UsernameException exception) {
         return new ResponseEntity<>(new ApiResponse(false, exception.getMessage()), HttpStatus.BAD_REQUEST);
