+++ backend/src/main/java/ubb/courses/backend/controllers/exceptions/user/UsernameException.java
@@ -1,4 +1,4 @@
-package ubb.courses.backend.controllers.exceptions.authentication;
+package ubb.courses.backend.controllers.exceptions.user;
 
 public class UsernameException extends RuntimeException {
     public UsernameException(String message) {
