+++ backend/src/main/java/ubb/courses/backend/services/UserService.java
@@ -6,8 +6,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import ubb.courses.backend.controllers.exceptions.AppException;
-import ubb.courses.backend.controllers.exceptions.authentication.EmailException;
-import ubb.courses.backend.controllers.exceptions.authentication.UsernameException;
+import ubb.courses.backend.controllers.exceptions.user.EmailException;
+import ubb.courses.backend.controllers.exceptions.user.UsernameException;
 import ubb.courses.backend.dtos.authentication.RegisterConverter;
 import ubb.courses.backend.models.authorization.Authority;
 import ubb.courses.backend.models.authorization.AuthorityType;
